backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.34.0
groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pydantic>=2.0.0
